tests/wpt/web-platform-tests/paint-timing/fcp-only/fcp-overflow.html

<!DOCTYPE html>
<head>
<title>Performance Paint Timing Test: FCP with element clipped with overflow</title>
<style>
    #main {
        width: 100px;
        height: 100px;
        overflow: hidden;
        display: none;
    }

    /* contentful class is defined in test_fcp script. */
    #main.contentful {
        display: block;
    }

    #text {
        top: 10000px;
        position: relative;
    }
</style>
</head>
<body>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../resources/utils.js"></script>
<div id="main">
    <div id="text">HELLO</div>
</div>
<script>
    test_fcp("First contentful paint fires even when element is hidden due to overflow.")
</script>
</body>

</html>
