third_party/blink/web_tests/external/wpt/paint-timing/fcp-only/fcp-overflow-expected.txt

This is a testharness.js-based test.
FAIL First contentful paint fires even when element is hidden due to overflow. assert_equals: Expected first contentful paint not found.  expected 1 but got 0
Harness: the test ran to completion.
